separating out tabs in account detail (#327)
